requirements.txt
# CAD Code-Based Generation - Requirements

# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Unsloth for efficient training (use specific versions)
unsloth==2025.3.18
unsloth_zoo==2025.3.16

# TRL for RLHF/DPO training
trl>=0.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector Search and Embeddings
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA is available
sentence-transformers>=2.2.0

# CAD/STEP File Processing
pythonocc-core>=7.7.0  # OpenCASCADE bindings
trimesh>=3.23.0
open3d>=0.17.0

# Visualization and Rendering
matplotlib>=3.7.0
Pillow>=10.0.0
imageio>=2.31.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Progress Bars
tqdm>=4.65.0

# Experiment Tracking
wandb>=0.15.0

# OpenAI API (for captioning)
openai>=1.0.0

# Tokenization
tiktoken>=0.5.0

# Utilities
jsonlines>=4.0.0
python-dotenv>=1.0.0

# Optional: Accelerate for distributed training
accelerate>=0.24.0

# Optional: PEFT for parameter-efficient fine-tuning
peft>=0.6.0

# Optional: BitsAndBytes for quantization
bitsandbytes>=0.41.0

# Note: Some dependencies may require specific CUDA versions
# For GPU acceleration:
# - Install faiss-gpu instead of faiss-cpu
# - Ensure PyTorch is installed with CUDA support
# - Check CUDA compatibility for bitsandbytes

# Installation notes:
# 1. Create conda environment: conda create -n cad_llm2 python=3.10
# 2. Activate environment: conda activate cad_llm2
# 3. Install PyTorch with CUDA: conda install pytorch torchvision torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
# 4. Install requirements: pip install -r requirements.txt
# 5. For pythonocc-core, you may need: conda install -c conda-forge pythonocc-core
